refactor: break down TUI app.rs into focused modules
- Split 466-line app.rs into organized module hierarchy
- All files under 200 lines (largest: 129 lines)
- Group by function: core, events, input, render, focus, tracing
- Maintain all functionality and compilation
