+++ crates/tui/src/app/mod.rs
@@ -1,0 +1,9 @@
+mod core;
+mod events;
+mod focus;
+mod input;
+mod render;
+mod tracing;
+
+pub use self::core::TuiApp;
+pub use self::tracing::tracing_to_tui_event;
